[#54]fix: 멘토링 피드백 반영
- Service 간 순환 참조 방지
- 패키지 구조 및 파일 위치 재조정
